requirements.txt
asttokens
certifi
charset-normalizer
colormath==3.0.0
comm
contourpy==1.2.1
cycler==0.12.1
debugpy
decorator
deltae2000==1.0.0
exceptiongroup
executing
filelock
flopth==0.1.5
fonttools==4.51.0
fsspec==2024.3.1
idna
imageio==2.34.1
importlib_metadata
ipykernel
ipython
jedi
Jinja2
jupyter_client
jupyter_core
kiwisolver==1.4.5
lazy_loader==0.4
MarkupSafe
matplotlib==3.8.4
matplotlib-inline
mkl-fft
mkl-random
mkl-service==2.4.0
mpmath
nest_asyncio
networkx
numpy
opencv-python==4.9.0.80
packaging
parso
pexpect
pickleshare
pillow
platformdirs
prompt-toolkit
psutil
ptyprocess
pure-eval
pyciede2000==0.0.21
Pygments
pyparsing==3.1.2
python-dateutil
PyYAML
pyzmq
requests
scikit-image==0.23.2
scipy==1.13.0
setuptools==69.5.1
six
stack-data
sympy
tabulate==0.9.0
thop==0.1.1.post2209072238
tifffile==2024.4.18
torchaudio==2.3.0
tornado
tqdm==4.66.2
traitlets
typing_extensions
urllib3
wcwidth
wheel==0.41.2
zipp